+++ src/lib.rs
@@ -260,7 +260,7 @@ impl<T: 'static, C: Config> Idr<T, C> {
     /// ```
     #[inline]
     pub fn get_owned(&self, key: Key) -> Option<OwnedEntry<T>> {
-        self.get(key, &EbrGuard::new()).map(BorrowedEntry::to_owned)
+        self.get(key, &EbrGuard::new())?.to_owned()
     }
 
     /// Returns `true` if the IDR contains an entry for the given key.